+++ src/main/java/com/vadimzu/webpcstore/model/Part.java
@@ -14,7 +14,7 @@ public class Part {
 
     private Long partID;
     private String partName;
-    private Long partPrice;
+    private Double partPrice;
     private Long stockQuantity;
     private String partType;
     private String shelf;
@@ -67,11 +67,11 @@ public class Part {
         this.stockQuantity = stockQuantity;
     }
 
-    public Long getPartPrice() {
+    public Double getPartPrice() {
         return partPrice;
     }
 
-    public void setPartPrice(Long partPrice) {
+    public void setPartPrice(Double partPrice) {
         this.partPrice = partPrice;
     }
